refactor: separate styles, elements directory with station_list
