+++ site/src/components/elements/station_list/styles.tsx
@@ -1,0 +1,7 @@
+import { styled } from '@junat/design'
+
+export const StyledStationList = styled('ul', {
+  display: 'flex',
+  flexDirection: 'column',
+  gap: '0.725rem'
+})
